+++ src/activitypub/handlers/create.rs
@@ -31,6 +31,7 @@ use mitra_utils::{
 use crate::activitypub::{
     constants::{AP_MEDIA_TYPE, AP_PUBLIC, AS_MEDIA_TYPE},
     deserialization::{
+        deserialize_into_object_id,
         parse_into_array,
         parse_into_href_array,
         parse_into_id_array,
@@ -694,6 +695,7 @@ pub async fn is_unsolicited_message(
 
 #[derive(Deserialize)]
 pub struct CreateNote {
+    #[serde(deserialize_with = "deserialize_into_object_id")]
     pub actor: String,
     pub object: Object,
 }